==> longer-version/1-dataprep-and-training-local.ipynb <==
import pandas as pd
import numpy as np

from sklearn.compose import make_column_selector as selector
from sklearn.preprocessing import OneHotEncoder, OrdinalEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import train_test_split

df = pd.read_csv('data/diabetic_data.csv')

df.head()

df.count()

uk_df = df.loc[: , (df == '?').any()]

df = df.replace('?', np.NaN) 

print(df.isna().any())

df.count()

df = df.drop(['weight'], axis=1)

df = df.dropna()

df.count()

df.loc[:, "medicare"] = (df.payer_code == "MC")
df.loc[:, "medicaid"] = (df.payer_code == "MD")

df.drop(['encounter_id', 'patient_nbr', 'payer_code', 'medical_specialty', 'admission_type_id', 
         'repaglinide','nateglinide','chlorpropamide','glimepiride','acetohexamide','glipizide','glyburide','tolbutamide',
        'pioglitazone','rosiglitazone','acarbose','miglitol','troglitazone','tolazamide','examide','citoglipton', 'metformin',
        'glyburide-metformin','glipizide-metformin','glimepiride-pioglitazone','metformin-rosiglitazone','metformin-pioglitazone', 'diag_2', 'diag_3', 'change'], axis=1, inplace=True)

df.loc[:, "age"] = df["age"].replace( ["[0-10)", "[10-20)", "[20-30)"], "30 years or younger")
df.loc[:, "age"] = df["age"].replace(["[30-40)", "[40-50)", "[50-60)"], "30-60 years")
df.loc[:, "age"] = df["age"].replace(["[60-70)", "[70-80)", "[80-90)", "[90-100)"], "Over 60 years")

df.rename(columns={ 'readmitted':'readmit_status', 'number_outpatient':'prior_outpatient', 'number_emergency': 'prior_emergency', 'number_inpatient':'prior_inpatient', 'diag_1':'primary_diagnosis', 'discharge_disposition_id':'discharge_destination', 'admission_source_id':'admission_source', 'diabetesMed':'diabetes_Med_prescribe'}, inplace=True)

df = df.drop(df[df.readmit_status ==  '>30'].index)
df['readmit_status'] = df['readmit_status'].replace({"NO":"not readmitted", "<30":"readmitted"})

    df.loc[:, "primary_diagnosis"] = df["primary_diagnosis"].replace(
        regex={
            "[7][1-3][0-9]": "Musculoskeltal primary_diagnosis",
            "250.*": "Diabetes",
            "[4][6-9][0-9]|[5][0-1][0-9]|786": "Respitory Issues",
            "[5][8-9][0-9]|[6][0-2][0-9]|788": "Genitourinary Issues"
        }
    )
    diagnoses = ["Respitory Issues", "Diabetes", "Genitourinary Issues", "Musculoskeltal Issues"]
    df.loc[:, "primary_diagnosis"] = df["primary_diagnosis"].apply(lambda x: x if x in diagnoses else "Other")

df = df[((df.discharge_destination != 11) & 
                                          (df.discharge_destination != 13) &
                                          (df.discharge_destination != 14) & 
                                          (df.discharge_destination != 19) & 
                                          (df.discharge_destination != 20) & 
                                          (df.discharge_destination != 21))] 

df.loc[:, "discharge_destination"] = (df.discharge_destination
                                            .apply(lambda x: "Discharged to Home" if x==1 else "Other"))

df.loc[:, "admission_source"] = df["admission_source"].apply(lambda x: x if x in ["Emergency", "Referral"] else "Other")

df = df.sample(frac=0.20)

df.count()

df.head()

df.dtypes


train, test = train_test_split(df, train_size=0.80, random_state=1)

train.to_parquet('data/training_data.parquet')
test.to_parquet('data/testing_data.parquet')

target_column_name = 'readmit_status'

def get_categorical_index(categorical_fields):
    cat_idx = []
    for col, value in categorical_fields.iteritems():
        if value.dtype == 'object':
            cat_idx.append(categorical_fields.columns.get_loc(col))
    print("col indices: ", cat_idx)  
    return cat_idx       

# Read in data
print("Reading data")
all_training_data = pd.read_parquet('data/training_data.parquet')
target = all_training_data[target_column_name]
features = all_training_data.drop([target_column_name], axis = 1)  

# Transform string data to numeric
numerical_selector = selector(dtype_include=np.number)
categorical_selector = selector(dtype_exclude=np.number)

numerical_columns = numerical_selector(features)
categorical_columns = categorical_selector(features)

categorial_encoder = OneHotEncoder(handle_unknown="ignore")
numerical_encoder = StandardScaler()

preprocessor = ColumnTransformer([
('categorical-encoder', categorial_encoder, categorical_columns),
('standard_scaler', numerical_encoder, numerical_columns)])

categorical_indices = get_categorical_index(features)
clf = make_pipeline(preprocessor, LogisticRegression())

X_train, X_test, y_train, y_test = train_test_split(features, target, 
test_size=0.3, random_state=1)

print("Training model...") 
model = clf.fit(X_train, y_train)
print("Accuracy score: ", clf.score(X_test,y_test))
